uplink: Config is now read before init - stack can now be selected
