+++ lib/uplink/config.hpp
@@ -1,0 +1,42 @@
+// This file is a part of the IncludeOS unikernel - www.includeos.org
+//
+// Copyright 2018 Oslo and Akershus University College of Applied Sciences
+// and Alfred Bratterud
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+#ifndef UPLINK_CONFIG_HPP
+#define UPLINK_CONFIG_HPP
+
+#include <string>
+#include <net/inet>
+#include <net/ip4/ip4.hpp>
+
+namespace uplink {
+
+  struct Config
+  {
+    net::Inet<net::IP4>* inet;
+    std::string url;
+    std::string token;
+    bool        reboot        = true;
+    bool        ws_logging    = true;
+    bool        serialize_ct  = false;
+
+    static Config read();
+  };
+
+}
+
+#endif
