+++ lib/uplink/starbase/vm.json
@@ -1,0 +1,7 @@
+{
+  "net" : [
+    {"device" : "vmxnet3", "mac" : "c0:01:0a:00:00:2a"},
+    {"device" : "vmxnet3", "mac" : "c0:01:0a:00:00:2b"},
+    {"device" : "vmxnet3", "mac" : "c0:01:0a:00:00:2c"}
+  ]
+}
