+++ lib/uplink/ws_uplink.cpp
@@ -39,12 +39,12 @@
 #include "log.hpp"
 
 namespace uplink {
-
-  const std::string WS_uplink::UPLINK_CFG_FILE{"config.json"};
   constexpr std::chrono::seconds WS_uplink::heartbeat_interval;
 
-  WS_uplink::WS_uplink(net::Inet<net::IP4>& inet)
-    : inet_{inet}, id_{inet.link_addr().to_string()},
+  WS_uplink::WS_uplink(Config config)
+    : config_{std::move(config)},
+      inet_{*config_.inet},
+      id_{inet_.link_addr().to_string()},
       parser_({this, &WS_uplink::handle_transport}),
       heartbeat_timer({this, &WS_uplink::on_heartbeat_timer})
   {
@@ -52,7 +52,6 @@ namespace uplink {
 
     liu::LiveUpdate::register_partition("uplink", {this, &WS_uplink::store});
 
-    read_config();
     CHECK(config_.reboot, "Reboot on panic");
 
     CHECK(config_.serialize_ct, "Serialize Conntrack");
@@ -61,7 +60,7 @@ namespace uplink {
 
     if(inet_.is_configured())
     {
-      start(inet);
+      start(inet_);
     }
     // if not, register on config event
     else
@@ -318,54 +317,6 @@ namespace uplink {
     [buffer] (auto) {
       liu::LiveUpdate::exec(buffer);
     });
-  }
-
-  void WS_uplink::read_config()
-  {
-    MYINFO("Reading uplink config");
-
-    const auto& cfg = ::Config::get();
-
-    Expects(not cfg.empty() && "Config is empty");
-
-    parse_config({cfg.data(), cfg.size()});
-  }
-
-  void WS_uplink::parse_config(const std::string& json)
-  {
-    using namespace rapidjson;
-    Document doc;
-    doc.Parse(json.data());
-
-    Expects(doc.IsObject() && "Malformed config");
-
-    Expects(doc.HasMember("uplink") && "Missing member \"uplink\"");
-
-    auto& cfg = doc["uplink"];
-
-    Expects(cfg.HasMember("url") && cfg.HasMember("token") && "Missing url or/and token");
-
-    config_.url   = cfg["url"].GetString();
-    config_.token = cfg["token"].GetString();
-
-    // Reboot on panic (optional)
-    if(cfg.HasMember("reboot"))
-    {
-      config_.reboot = cfg["reboot"].GetBool();
-    }
-
-    // Log over websocket (optional)
-    if(cfg.HasMember("ws_logging"))
-    {
-      config_.ws_logging = cfg["ws_logging"].GetBool();
-    }
-
-    // Serialize conntrack
-    if(cfg.HasMember("serialize_ct"))
-    {
-      config_.serialize_ct = cfg["serialize_ct"].GetBool();
-    }
-
   }
 
   template <typename Writer, typename Stack_ptr>
